Add interface namespace to HubClientToGenerate
A new attribute, 'InterfaceNamespace', was added to the 'HubClientToGenerate' class to hold the namespace information for the interface. This namespace is now included among the usings in 'HubClientSource'. Changes were also made to 'SignalRClientGenerator' to get the namespace of the current interface being processed. The test cases were updated to reflect these changes.