Add performance improvements for Source Gen (#10)
* Refactor HubClientToGenerate to a record type

In the SignalRGen.Generator, the HubClientToGenerate class has been refactored to a record type in HubClientToGenerate.cs. This has simplified the code and removed unnecessary methods. The SignalRClientGenerator and UnitTestSessionStore identifier in .DotSettings.user file have also been updated accordingly.

* WIP switching to cacheable types

This commit is partway through the conversion of the Source Gen to use cacheable types.
Methods are already switched to use a cacheable type structure, usings are still not cacheable and will be done in a follow-up commit.

* Refactor code to use CacheableUsingDeclaration

The commit changes the data structure used for storing 'using' directives in the 'HubClientToGenerate' class and `SignalRClientGenerator` method `GetInterfacesUsings`. It replaces 'UsingDirectiveSyntax' with 'CacheableUsingDeclaration' to more efficiently cache and retrieve 'using' directive information across instances.

* Update HubClientSource and test to handle multiple parameters

The methods HubClientTests.Generates_HubClient_Correctly and HubClientSource have been updated to handle multiple method parameters. Also, the 'Microsoft.AspNetCore.SignalR.Client' namespace is now included among the usings.

* Add interface namespace to HubClientToGenerate

A new attribute, 'InterfaceNamespace', was added to the 'HubClientToGenerate' class to hold the namespace information for the interface. This namespace is now included among the usings in 'HubClientSource'. Changes were also made to 'SignalRClientGenerator' to get the namespace of the current interface being processed. The test cases were updated to reflect these changes.

* Add comments for error analysis and future fixes in SignalRClientGenerator

In the SignalRClientGenerator.cs file, comments were added to guide future work.

* Update package versions and fix tests

Multiple package versions have been updated to their newer versions to keep the project up-to-date. Additionally, the tests missed the assembly for the default .Net libraries like `System`.

* Enhance source generator testing with cacheability checks

In this commit, we extend the `TestHelper` class in the `SignalRGen.Generator.Tests` project to include cacheability assertion. This addition aims to verify that source generator outputs remain consistent across different compilation states and are safely cacheable. The `SignalRClientGenerator.cs` file is updated to use specific tracking names for phases in source generation. A new file `TrackingNames.cs` has been added to define these constant string tracking names. The inclusion of the `FluentAssertions` library facilitates the verification process.