Add support for server-to-client and client-to-server methods (#16)
* Add server-to-client and client-to-server attributes

Generated two new attributes, ServerToClientMethodAttribute and ClientToServerMethodAttribute. These attributes will enable more specificity and control in method generation. The under-the-hood logic was adjusted to recognize and apply these attributes during generation. The client-to-server generation is not yet done.

* Fix equality check recursion resulting in a StackOverflow exception

The Equals method in the EquatableArray class has been updated. The logic of the object comparison has been optimized to not indefinitely recurse into a StackOverflow exception.

* Add client-to-server support

Updated the structure of the SignalRClientGenerator code and included support for client-to-server method generation.

* Fix generic return type for client-to-server method invocations

The code has been adjusted to correctly declare the generic return type when invoking async methods in the SignalR Client Hub classes. This commit also includes the adjustment of method indentation for better readability.

* Fix ValidateHubConnection method placement being outside HubClient class

The ValidateHubConnection method was moved inside the last bracket in the HubClient tests and the main source file. With the new positioning, it remains within the same class scope.

* Update package versions in multiple project files

Updated the versions of Microsoft.NET.Test.Sdk, xunit, xunit.runner.visualstudio, and Swashbuckle.AspNetCore in the test and example project files. Additionally, a new package reference to Microsoft.Net.Compilers.Toolset was added to the Example.Contracts project for building on machines with older .Net SDK versions.

* Add new communication methods to SignalR hub

The commit introduces two new communication methods to the SignalR hub: a method that sends a message and returns a string, and a method that sends a message without returning a value. This change also includes modification in the Blazor client to invoke the SendExampleMessage function.

* Add summary comments to generated client-to-server HubClient methods

Introduced summary comments in the SignalRGen.Generated HubClient methods. These comments explain the purpose of each generated method, facilitating easier usage and understanding of the code. The changes affect both test files and the HubClientSource.cs.

* Add exception description to HubClient methods

The code has been updated to include a more descriptive exception documentation on HubClient methods. It now specifies that InvalidOperationException might be thrown when a Hub isn't started by calling TestHubClient.StartAsync. The template in HubClientSource.cs was adjusted to include this information as well.