Upgrade projects to target .NET 9.0. (#29)
* Upgrade projects to target .NET 9.0 framework.

Updated the TargetFramework property in multiple csproj files from .NET 8.0 to .NET 9.0 to align with the newer framework version. This improves compatibility with the latest features and ensures forward-looking support for the projects.

* Update CI to use .NET 9.0.x

Bump the .NET version in the GitHub Actions workflow from 8.0.x to 9.0.x. Ensures the CI pipeline uses the latest version for compatibility and new features.