Extract static files into common project (#73)
* Add SignalRGen.Abstractions project with initial setup

* Remove `IHubClient` and migrate SignalRGen.Abstractions to .NET Standard 2.0 with added compiler support

Ditched the `IHubClient` due to abstract static fields in Interfaces not supported by netstandard2.0. The `HubUri` that the `IHubClient` exposed needs to be generated by the source gen from now on for each `HubClient`.

* Migrate attribute definitions and options to SignalRGen.Abstractions with namespace adjustments

* Fix typo in source file name for `HubClientBase.g.cs` generation

* Refactor tests for common file extraction

Refactored the tests to adhere to the extraction of the common files into the Abstraction project.

* Cleanup